rm old conversions, refactor some utilities out of DartsAiModel to make it a bit leaner.
