--- src/main/resources/sql/v10/1. DartzeeRule.sql
@@ -1,3 +1,0 @@
-ALTER TABLE DartzeeRule ALTER COLUMN Dart1Rule SET DATA TYPE VARCHAR(32000);
-ALTER TABLE DartzeeRule ALTER COLUMN Dart2Rule SET DATA TYPE VARCHAR(32000);
-ALTER TABLE DartzeeRule ALTER COLUMN Dart3Rule SET DATA TYPE VARCHAR(32000)
