--- src/main/resources/sql/v10/2. Game.sql
@@ -1,6 +1,0 @@
-UPDATE
-    GAME
-SET
-    MATCHORDINAL = MATCHORDINAL + 1
-WHERE
-    DARTSMATCHID IN (SELECT DartsMatcHId FROM Game WHERE MatchOrdinal = 0)
