+++ src/main/resources/sql/v13/1. Player.sql
@@ -1,0 +1,1 @@
+ALTER TABLE Player DROP COLUMN Strategy;
